+++ src/Service/Rekognition/src/RekognitionClient.php
@@ -228,7 +228,7 @@ class RekognitionClient extends AbstractApi
     }
 
     /**
-     * Gets the name and additional information about a celebrity based on his or her Amazon Rekognition ID. The additional
+     * Gets the name and additional information about a celebrity based on their Amazon Rekognition ID. The additional
      * information is returned as an array of URLs. If there is no additional information about the celebrity, this list is
      * empty.
      *