+++ src/Service/Rekognition/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - AWS api-change: Add AWS tagging support for Amazon Rekognition collections, stream processors, and Custom Label models
 - AWS api-change: Improve documentation
+- AWS api-change: Add new attributes to Rekognition RecognizeCelebities and GetCelebrityInfo API operations.
 
 ### Fixed